webapi/tct-backgrounds-css3-tests/backgrounds/CSS3BG_border-image-repeat_repeat-x_url_png_auto.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu, Jinfeng <jinfengx.liu@intel.com>

-->

<html>
  <head>
    <title>CSS3 Backgrounds Test: CSS3BG_border-image-repeat_repeat-x_url_png_auto</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-css3-background-20120724/#border-image-repeat" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if 'border-image-repeat:repeat-x;border-image-source:url(support/swatch-orange.png);border-image-width:auto' on test div" />
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
    <script type="text/javascript" src="support/support.js"></script>
    <style>
        #test{
            height: 200px;
            width: 300px;
            padding: 5px;
            border: 5px solid black;
            margin: 5px;
            background: blue;
        }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script type="text/javascript">
        var div = document.querySelector("#test");
        var t = async_test(document.title, {timeout: 500});
        t.step(function () {
            div.style[headProp("border-image-repeat")] = "repeat-x";
            div.style[headProp("border-image-source")] = "url(support/swatch-orange.png)";
            div.style[headProp("border-image-width")] = "auto";
            var propvalue = GetCurrentStyle("border-image-repeat");
            assert_equals(propvalue, "stretch", "The element border-image-repeat should ");
            var borderImageWidth = GetCurrentStyle("border-image-width");
            assert_equals(borderImageWidth, "auto", "The element border-image-width should ");
            var borderImageSource = GetCurrentStyle("border-image-source");
            prop = borderImageSource.indexOf("support/swatch-orange.png") != -1;
            assert_true(prop, "The element border-image-source should be url(support/swatch-orange.png)");
        });
        t.done();
    </script>
  </body>
</html>
